# Fix packed enum decoding benchmark